WIP: Refactor mobile phone updating process
