+++ app/controllers/account/activations_controller.rb
@@ -2,6 +2,7 @@
 
 module Account
   class ActivationsController < ApplicationController
+    include PhoneVerification
     before_action :redirect_to_root, if: -> { user_signed_in? }
     skip_before_action :authenticate_user!
     layout 'guest'
@@ -18,36 +19,11 @@ module Account
       respond_to :js
     end
 
-    # rubocop:disable Metrics/AbcSize
     def check_phone_verification
-      verify = params[:phone_verification]
-      response = Twilio::Verify.check_verification_code(verify[:mobile_phone], verify[:verification_code])
-
-      if response == 'ok'
-        return redirect_to login_path, notice: t('.success') if update_process(session[:user_id],
-                                                                               verify[:mobile_phone])
-
-        redirect_to activation_path, alert: t('errors.system_error')
-      else
-        redirect_to activation_path,
-                    alert: t("twilio.errors.#{response.error_code}", default: t('errors.system_error'))
-      end
+      check(params, login_path, activation_path)
     end
-    # rubocop:enable Metrics/AbcSize
 
     private
-
-    def update_process(user_id, phone)
-      user = User.find(user_id)
-      user.transaction do
-        user.prospective_students.registered.update(archived: true)
-        user.prospective_employees.update(archived: true)
-        user.update!(mobile_phone: phone, activated: true, activated_at: Time.zone.now)
-      end
-    rescue ActiveRecord::RecordNotFound, ActiveRecord::RecordInvalid => e
-      Rollbar.error(e, e.message)
-      false
-    end
 
     def redirect_to_root
       redirect_to(:root)
