+++ app/views/account/activations/create.js.erb
@@ -7,6 +7,9 @@ $('#error_messages').html('');
     )
   <% end %>
 <% else %>
-  $('#verification_modal').html("<%= escape_javascript(render 'modal') %>");
+  $('#verification_modal').html("<%= escape_javascript(render partial: 'layouts/shared/phone_verification_modal', \
+                                                              locals: { action: 'check_phone_verification', \
+                                                                        controller: 'account/activations', \
+                                                                        mobile_phone: @activation.mobile_phone }) %>");
   $('#verification_modal').modal('show')
 <% end %>
