+++ config/locales/controllers/account/registrations.en.yml
@@ -14,9 +14,5 @@ en:
         current_password_hint: you must enter your current password for changing your e-mail address
         leave_blank: leave blank if you don't want to change
         mobile_phone_updating: Mobile Phone Updating
-      modal:
-        verification_code: Verification Code
-        verify: Verify
-        verify_your_phone: Verify Your Phone
       phone_verification:
         cannot_change_phone: You cannot make a mobile phone update request with the same number
