+++ config/locales/controllers/account/registrations.tr.yml
@@ -14,9 +14,5 @@ tr:
         current_password_hint: e-mail adresinizi değiştirebilmek için mevcut parolanızı girmeniz gerekmektedir
         leave_blank: değiştirmek istemiyorsanız boş bırakın
         mobile_phone_updating: Cep Telefonu Güncelleme
-      modal:
-        verification_code: Doğrulama Kodu
-        verify: Doğrula
-        verify_your_phone: Telefonunuzu Doğrulayın
       phone_verification:
         cannot_change_phone: Telefon güncelleme talebinizi aynı telefon numarası ile yapamazsınız.
